Adds the functionality to notify Actors of unhandled exceptions (#48)
## What is the goal of this PR?

Currently, there is a major limitation in the Actor library such that an unhandled exception in an actor instance would be swallowed by the event loop.

This PR adds the functionality allowing actors to be notified of the exception that it failed to handle. This would allow actors to decide on whether to recover (if said actor considers the exception non-fatal) or shut down (if it considers the exception fatal).


```
class MyActor extends Actor<MyActor> {
  @Override
  public void exception(Exception e) {
    // handle the uncaught exception
  }
}
```

This PR depends on: https://github.com/graknlabs/common/pull/47.