When running tests, poll used ports to identify when Grakn starts (#88)
## What is the goal of this PR?

Previously we waited N milliseconds and then ran the tests. This was wasteful, because often Grakn would start way in advance; and unreliable, because sometimes Grakn would not start in time and CI would fail. Now, we add a polling loop that checks if the server port is in use, and signals that the tests may begin once the port is in use.

## What are the changes implemented in this PR?

- When setting up tests that use a background Grakn server, poll the used ports every 500ms to detect when Grakn is alive, and then run the tests