Fix runner archive extraction error (#128)
## What is the goal of this PR?

TypeDB Runner now handles unarchiving artifacts in a more rigorous way with respect to the provided indices. Now, we correctly check that a provided index is valid prior to an array lookup.

## What are the changes implemented in this PR?

Prior to unarchiving a runner we need to get its path from the arguments passed to the JVM. We want to ensure that index is strictly less than args.length so that
```java
File file = new File(args[index]);
```
is valid.

We check if index is equal to args.length as well as greater than, as both will lead to the `args[index]` call to fail.