Include TypeDB runners in common library jar (#139)
## What is the goal of this PR?

Currently, we don't include the test directory files in the main jar
produced by typedb-common. As raised by a community member, we should
include this so other people can make use of this via the provided jars.

## What are the changes implemented in this PR?

We've included all of the java files found under //test in the `srcs`
argument to `java_library`.

We've also corrected one scary \*/\* import and now explicitly specify
what directory is being brought in.