Changed float to double in YAML parser (#143)
## What is the goal of this PR?

The float parsing in the YAML parser has been replaced with double
parsing. This is because Java attempts to parse float-like YAML strings
to double and throws an error when parsing them to float, and float
parsing hadn't been used anywhere so it went unnoticed. Float-like
strings are now parsed as doubles.

## What are the changes implemented in this PR?

All references to floats in the YAML class have been replaced with
doubles in class names, method names, etc.