Improve entry point runner UX (#158)
## What is the goal of this PR?

We provide a single `server` entry point for starting up both core and
enterprise server, whichever happens to be installed. We also omit
`server` or `console` from the usage help if it is not present to avoid
confusion. See https://github.com/vaticle/typedb/issues/6942 for more
details.

We also improve the UX of the windows version of the entry point.
Console no longer opens in a new window, `--help` is printed inline as
expected, and in the event of the server failure, the logs are displayed
to the user.