Fix line endings for TypeDB batch launcher (#162)
## What is the goal of this PR?

When launching the server, a bug caused additional output that claimed a
batch label that did exist, didn't. We've corrected the line endings
from LF to CRLF on TypeDB's batch launcher.

## What are the changes implemented in this PR?

As above.
* Changed the line endings on typedb.bat.
* Added a gitattributes file to specify the above requirement for bat
files for the future.
* Added checkstyle for the above new file.
* Added an end of line and fixed a space.

## Notes

The line endings change seems to be invisible to git and GitHub's
changes viewer. To verify this change, checkout and run `cat -ev
binary/typedb.bat` to see the new line endings.